Refactor and add viewport math
