+++ src/d3/mod.rs
@@ -1,0 +1,1 @@
+pub mod scale;
