+++ src/d3/scale.rs
