+++ src/plots/mod.rs
@@ -1,0 +1,3 @@
+pub mod triangle;
+pub mod scatterplot;
+pub mod bioimage;
