+++ src/plots/shaders/scatterplot.vs.wgsl
